+++ backend/src/test/java/tech/geocodeapp/geocode/Collectable_temp/CollectableServiceImplTest.java
@@ -1,4 +1,4 @@
-package tech.geocodeapp.geocode.collectable;
+package tech.geocodeapp.geocode.Collectable_temp;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
