+++ backend/src/test/java/tech/geocodeapp/geocode/Collectable_temp/CollectableSetMockRepository.java
@@ -1,4 +1,4 @@
-package tech.geocodeapp.geocode.collectable;
+package tech.geocodeapp.geocode.Collectable_temp;
 
 import tech.geocodeapp.geocode.collectable.model.CollectableSet;
 import org.springframework.data.domain.Example;
