+++ backend/src/test/java/tech/geocodeapp/geocode/GeoCode_temp/GeoCodeServiceImplTest.java
@@ -1,4 +1,4 @@
-package tech.geocodeapp.geocode.geocode;
+package tech.geocodeapp.geocode.GeoCode_temp;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.mockito.Mock;
