+++ backend/src/test/java/tech/geocodeapp/geocode/User_temp/UserMockRepository.java
@@ -1,4 +1,4 @@
-package tech.geocodeapp.geocode.user;
+package tech.geocodeapp.geocode.User_temp;
 
 import org.springframework.data.domain.Example;
 import org.springframework.data.domain.Page;
