+++ backend/src/test/java/tech/geocodeapp/geocode/User_temp/UserServiceImplTest.java
@@ -1,39 +1,58 @@
-package tech.geocodeapp.geocode.user;
+package tech.geocodeapp.geocode.User_temp;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
-import org.junit.jupiter.api.*;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-
+import tech.geocodeapp.geocode.Collectable_temp.CollectableMockRepository;
+import tech.geocodeapp.geocode.Collectable_temp.CollectableTypeMockRepository;
 import tech.geocodeapp.geocode.collectable.model.Collectable;
-import tech.geocodeapp.geocode.collectable.repository.CollectableTypeRepository;
+import tech.geocodeapp.geocode.collectable.model.CollectableType;
+import tech.geocodeapp.geocode.GeoCode_temp.GeoCodeMockRepository;
 import tech.geocodeapp.geocode.user.exception.NullUserRequestParameterException;
-import tech.geocodeapp.geocode.user.repository.UserRepository;
 import tech.geocodeapp.geocode.user.service.*;
 import tech.geocodeapp.geocode.user.request.*;
 import tech.geocodeapp.geocode.user.response.*;
 
-@SpringBootTest
-@TestMethodOrder( MethodOrderer.OrderAnnotation.class )
-public class UserServiceImplIT {
-    @Autowired
+
+@ExtendWith( MockitoExtension.class )
+public class UserServiceImplTest {
     private UserService userService;
-
-    @Autowired
-    private UserRepository userRepo;
-
-    @Autowired
-    private CollectableTypeRepository collectableTypeRepo;
 
     private final UUID invalidUserId = UUID.fromString("31d72621-091c-49ad-9c28-8abda8b8f055");
     private final UUID validUserId = UUID.fromString("183e06b6-2130-45e3-8b43-634ccd3e8e6f");
     private final String invalidUserIdMessage = "Invalid user id";
+    private GetCurrentCollectableResponse getCurrentCollectableResponse;
+    private GetUserTrackableResponse getUserTrackableResponse;
+    private UpdateLocationResponse updateLocationResponse;
+
+    UserServiceImplTest() {
+
+    }
+
+    @BeforeEach
+    void setup() {
+        CollectableTypeMockRepository collectableTypeMockRepo = new CollectableTypeMockRepository();
+        userService = new UserServiceImpl( new UserMockRepository(), new CollectableMockRepository(), collectableTypeMockRepo, new GeoCodeMockRepository());
+
+        //save the valid trackable CollectableType
+        CollectableType trackableCollectableType = new CollectableType();
+        trackableCollectableType.setId(UUID.fromString("0855b7da-bdad-44b7-9c22-18fe266ceaf3"));
+        collectableTypeMockRepo.save(trackableCollectableType);
+
+        //save the valid user to the MockRepo
+        userService.registerNewUser(validUserId, "john_smith");
+    }
 
     @Test
-    public void getCurrentCollectableTestNullRequest() throws NullUserRequestParameterException {
+    public void getCurrentCollectableTestNullRequest() {
         try{
             GetCurrentCollectableResponse response = userService.getCurrentCollectable(null);
             Assertions.assertFalse(response.isSuccess());
@@ -42,10 +61,20 @@ public class UserServiceImplIT {
         }catch (NullUserRequestParameterException e){
             Assertions.fail(e.getMessage());
         }
-    }//
+    }
 
     @Test
-    public void getCurrentCollectableTestInvalidUser() throws NullUserRequestParameterException {
+    public void getCurrentCollectableTestNullId(){
+           GetCurrentCollectableRequest request = new GetCurrentCollectableRequest();
+           request.setUserID(null);
+
+           //GetCurrentCollectableResponse response = userService.getCurrentCollectable(request);
+           assertThatThrownBy(() -> getCurrentCollectableResponse = userService.getCurrentCollectable(request))
+                .isInstanceOf(NullUserRequestParameterException.class);
+    }
+
+    @Test
+    public void getCurrentCollectableTestInvalidUser() {
         try{
             /*
             Create a request object
@@ -61,10 +90,10 @@ public class UserServiceImplIT {
         }catch (NullUserRequestParameterException e){
             Assertions.fail(e.getMessage());
         }
-    }//
+    }
 
     @Test
-    public void getCurrentCollectableTestValidUser() throws Exception {
+    public void getCurrentCollectableTestValidUser() {
         try{
             /*
             Create a request object
@@ -83,7 +112,7 @@ public class UserServiceImplIT {
     }
 
     @Test
-    public void getUserTrackableTestNullRequest() throws NullUserRequestParameterException {
+    public void getUserTrackableTestNullRequest() {
         try{
             GetUserTrackableResponse response = userService.getUserTrackable(null);
             Assertions.assertFalse(response.isSuccess());
@@ -95,7 +124,16 @@ public class UserServiceImplIT {
     }
 
     @Test
-    public void getUserTrackableTestInvalidUser() throws NullUserRequestParameterException {
+    public void getUserTrackableTestNullId(){
+        GetUserTrackableRequest request = new GetUserTrackableRequest();
+        request.setUserID(null);
+
+        assertThatThrownBy(() -> getUserTrackableResponse = userService.getUserTrackable(request))
+                .isInstanceOf(NullUserRequestParameterException.class);
+    }
+
+    @Test
+    public void getUserTrackableTestInvalidUser() {
         try{
             /*
             Create a request object
@@ -114,7 +152,7 @@ public class UserServiceImplIT {
     }
 
     @Test
-    public void getUserTrackableTestValidUser() throws NullUserRequestParameterException {
+    public void getUserTrackableTestValidUser() {
         try{
             /*
             Create a request object
@@ -136,7 +174,7 @@ public class UserServiceImplIT {
     }
 
     @Test
-    public void updateLocationTestNullRequest() throws NullUserRequestParameterException {
+    public void updateLocationTestNullRequest() {
         try{
             UpdateLocationResponse response = userService.updateLocation(null);
             Assertions.assertFalse(response.isSuccess());
@@ -148,7 +186,16 @@ public class UserServiceImplIT {
     }
 
     @Test
-    public void updateLocationTestInvalidUser() throws NullUserRequestParameterException {
+    public void updateLocationTestNullId(){
+        UpdateLocationRequest request = new UpdateLocationRequest();
+        request.setUserID(null);
+
+        assertThatThrownBy(() -> updateLocationResponse = userService.updateLocation(request))
+                .isInstanceOf(NullUserRequestParameterException.class);
+    }
+
+    @Test
+    public void updateLocationTestInvalidUser() {
         try{
             /*
             Create a request object
@@ -167,7 +214,7 @@ public class UserServiceImplIT {
     }
 
     @Test
-    public void updateLocationTestValidUser() throws NullUserRequestParameterException {
+    public void updateLocationTestValidUser() {
         try{
             /*
              Create a request object
